* 'master' of github.com:daitss/core:
  Package#normal_events omits noisy events
  ability to unstash all packages in a bin
  direct submission client, resolves #275
  sip_archive now supports directories, fix bug where account/project missing from sip descriptor
  fix bug
  patch for #255

+++ lib/daitss/proc/wip/process.rb
@@ -38,7 +38,7 @@ module Daitss
     def stash_journal
       jf = File.join @path, 'journal'
       pf = profile_file('journal')
-      FileUtils.cp jf, pf
+      FileUtils.cp jf, pf if File.exist? jf
     end
 
     def procname
